bubble shield + spikes
spikes are now added along with ring dropping, sonic will now drop his rings upon taking damage. if he has none, he will die.

the bubble shield is here, it is not the most accurate since i was not able to get the bubble shield formula working with godot, but it still works.

ring sounds also now play from the player so that if you collect a bunch of rings at once, it wont be very loud

camera also now pans when looking up and down. this was done using the offset of the camera.